Ahora filtrar y buscar producto regresan los productos en lugar de solo el id
